refactore code

+++ Comment/views.py
@@ -18,7 +18,6 @@ class CommentsPostlist(generics.RetrieveAPIView):
 
 @api_view(['GET', 'POST'])
 def commentslist(request, id_post):
-    print('entou')
     try:
         post = Post.objects.get(id=id_post)
     except Post.DoesNotExist:
@@ -32,13 +31,17 @@ def commentslist(request, id_post):
         if request.user.is_authenticated:
             comment_serializer = CommentsSerializer(Comment, data=request.data)
             if comment_serializer.is_valid():
-                comment = Comment(name=request.data['name'], email=request.data['email'], body=request.data['body'],
+                comment = Comment(name=request.data['name'],
+                                  email=request.data['email'],
+                                  body=request.data['body'],
                                   postId=post)
                 comment.save()
                 return Response(request.data)
-            return Response(comment_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response(comment_serializer.errors,
+                            status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response({'User not authenticated'}, status=status.HTTP_401_UNAUTHORIZED)
+            return Response({'User not authenticated'},
+                            status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['GET', 'PUT', 'DELETE'])
@@ -57,24 +60,31 @@ def comments_post_detail(request, id_post, id_comment):
         return Response(comment_serializer.data)
 
     elif request.method == 'PUT':
-        comment_serializer = CommentsUpdateSerializer(Comment, data=request.data)
-        if comment_serializer.is_valid():
-            comment.name = request.data['name']
-            comment.email = request.data['email']
-            comment.body = request.data['body']
-            comment.save()
-            return Response(request.data)
-        return Response(comment_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
+            user = request.user
+            post = Post.objects.get(id=id_post)
+            if post.owner == user:
+                comment_serializer = CommentsUpdateSerializer(Comment, data=request.data)
+                if comment_serializer.is_valid():
+                    comment.name = request.data['name']
+                    comment.email = request.data['email']
+                    comment.body = request.data['body']
+                    comment.save()
+                    return Response(request.data, status=status.HTTP_200_OK)
+                return Response(comment_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            else:
+                return Response({'No permission / UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
+        else:
+            return Response({'User not authenticated'}, status=status.HTTP_401_UNAUTHORIZED)
 
     elif request.method == 'DELETE':
-        user = request.user
-        post = Post.objects.get(id=id_post)
-        if post.owner == user:
-            comment.delete()
-            return Response({'result': 'comment deleted'}, status=status.HTTP_204_NO_CONTENT)
+        if request.user.is_authenticated:
+            user = request.user
+            post = Post.objects.get(id=id_post)
+            if post.owner == user:
+                comment.delete()
+                return Response({'result': 'comment deleted'}, status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response({'No permission / UNAUTHORIZED'}, status=status.HTTP_401_UNAUTHORIZED)
         else:
-            return Response({'Você não tem autorização para deletar o comentario'}, status=status.HTTP_401_UNAUTHORIZED)
-
-
-
-
+            return Response({'User not authenticated'}, status=status.HTTP_401_UNAUTHORIZED)
